Move vector normalizers to o.l.transform.normalize
