+++ lenskit-core/src/main/java/org/lenskit/transform/normalize/AbstractItemVectorNormalizer.java
@@ -18,7 +18,7 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
-package org.grouplens.lenskit.transform.normalize;
+package org.lenskit.transform.normalize;
 
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.grouplens.lenskit.vectors.SparseVector;
@@ -27,12 +27,9 @@ import javax.annotation.Nonnull;
 import javax.annotation.Nullable;
 
 /**
- * Abstract user vector normalizer implementation.
- *
- * @author <a href="http://www.grouplens.org">GroupLens Research</a>
- * @since 0.11
+ * Abstract item vector normalizer implementation.
  */
-public abstract class AbstractUserVectorNormalizer implements UserVectorNormalizer {
+public abstract class AbstractItemVectorNormalizer implements ItemVectorNormalizer {
 
     /**
      * {@inheritDoc}
@@ -40,14 +37,14 @@ public abstract class AbstractUserVectorNormalizer implements UserVectorNormaliz
      * and the resulting {@link VectorTransformation}.
      */
     @Override
-    public MutableSparseVector normalize(long user, @Nonnull SparseVector vector,
+    public MutableSparseVector normalize(long item, @Nonnull SparseVector vector,
                                          @Nullable MutableSparseVector target) {
         MutableSparseVector v = target;
         if (v == null) {
             v = vector.mutableCopy();
         }
 
-        VectorTransformation tform = makeTransformation(user, vector);
+        VectorTransformation tform = makeTransformation(item, vector);
         return tform.apply(v);
     }
 }
