+++ lenskit-core/src/main/java/org/lenskit/transform/normalize/AbstractVectorNormalizer.java
@@ -18,8 +18,35 @@
  * this program; if not, write to the Free Software Foundation, Inc., 51
  * Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
  */
+package org.lenskit.transform.normalize;
+
+import org.grouplens.lenskit.vectors.MutableSparseVector;
+import org.grouplens.lenskit.vectors.SparseVector;
+
+import javax.annotation.Nonnull;
+import javax.annotation.Nullable;
+
 /**
- * Interfaces and classes for data normalization. Normalization is typically
- * done prior to training a recommender.
+ * Abstract vector normalizer implementation.
+ *
+ * @author <a href="http://www.grouplens.org">GroupLens Research</a>
  */
-package org.grouplens.lenskit.transform.normalize;
+public abstract class AbstractVectorNormalizer implements VectorNormalizer {
+
+    /**
+     * {@inheritDoc}
+     * <p>Delegates to {@link #makeTransformation(SparseVector)} and the
+     * resulting {@link VectorTransformation}.
+     */
+    @Override
+    public MutableSparseVector normalize(@Nonnull SparseVector reference,
+                                         @Nullable MutableSparseVector target) {
+        MutableSparseVector v = target;
+        if (v == null) {
+            v = reference.mutableCopy();
+        }
+
+        VectorTransformation tform = makeTransformation(reference);
+        return tform.apply(v);
+    }
+}
