+++ lenskit-integration-tests/src/it/gradle/user-user-identical/algorithms.groovy
@@ -21,9 +21,9 @@
 
 
 import org.grouplens.lenskit.knn.item.model.*
-import org.grouplens.lenskit.transform.normalize.MeanCenteringVectorNormalizer
-import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer
-import org.grouplens.lenskit.transform.normalize.VectorNormalizer
+import org.lenskit.transform.normalize.MeanCenteringVectorNormalizer
+import org.lenskit.transform.normalize.UserVectorNormalizer
+import org.lenskit.transform.normalize.VectorNormalizer
 import org.lenskit.api.ItemScorer
 import org.lenskit.knn.NeighborhoodSize
 import org.lenskit.knn.user.NeighborFinder
