+++ lenskit-knn/src/main/java/org/lenskit/knn/item/ItemItemScorer.java
@@ -25,8 +25,7 @@ import it.unimi.dsi.fastutil.longs.Long2DoubleOpenHashMap;
 import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.fastutil.longs.LongIterators;
 import org.grouplens.lenskit.symbols.Symbol;
-import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer;
-import org.grouplens.lenskit.transform.normalize.VectorTransformation;
+import org.lenskit.transform.normalize.UserVectorNormalizer;
 import org.lenskit.util.ScoredIdAccumulator;
 import org.lenskit.util.TopNScoredIdAccumulator;
 import org.lenskit.util.UnlimitedScoredIdAccumulator;
