+++ lenskit-knn/src/main/java/org/lenskit/knn/item/model/ItemItemBuildContextProvider.java
@@ -26,13 +26,13 @@ import it.unimi.dsi.fastutil.longs.Long2ObjectOpenHashMap;
 import it.unimi.dsi.fastutil.longs.LongSortedSet;
 import org.grouplens.lenskit.scored.ScoredIdListBuilder;
 import org.grouplens.lenskit.scored.ScoredIds;
-import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer;
 import org.grouplens.lenskit.vectors.ImmutableSparseVector;
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.grouplens.lenskit.vectors.SparseVector;
 import org.grouplens.lenskit.vectors.VectorEntry;
 import org.lenskit.data.ratings.RatingVectorPDAO;
 import org.lenskit.inject.Transient;
+import org.lenskit.transform.normalize.UserVectorNormalizer;
 import org.lenskit.util.IdBox;
 import org.lenskit.util.collections.LongUtils;
 import org.lenskit.util.io.ObjectStream;
