+++ lenskit-knn/src/main/java/org/lenskit/knn/item/model/NormalizingItemItemModelProvider.java
@@ -26,12 +26,12 @@ import com.google.common.collect.Lists;
 import it.unimi.dsi.fastutil.longs.Long2DoubleMap;
 import it.unimi.dsi.fastutil.longs.LongIterator;
 import it.unimi.dsi.fastutil.longs.LongSortedSet;
-import org.grouplens.lenskit.transform.normalize.ItemVectorNormalizer;
 import org.grouplens.lenskit.transform.truncate.VectorTruncator;
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.grouplens.lenskit.vectors.SparseVector;
 import org.lenskit.inject.Transient;
 import org.lenskit.knn.item.ItemSimilarity;
+import org.lenskit.transform.normalize.ItemVectorNormalizer;
 import org.lenskit.util.collections.LongUtils;
 import org.lenskit.util.keys.SortedKeyIndex;
 import org.slf4j.Logger;
