+++ lenskit-knn/src/main/java/org/lenskit/knn/user/UserSnapshot.java
@@ -24,12 +24,12 @@ import com.google.common.base.Preconditions;
 import com.google.common.collect.ImmutableList;
 import it.unimi.dsi.fastutil.longs.*;
 import org.grouplens.grapht.annotation.DefaultProvider;
-import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer;
 import org.grouplens.lenskit.vectors.ImmutableSparseVector;
 import org.grouplens.lenskit.vectors.MutableSparseVector;
 import org.lenskit.data.ratings.RatingVectorPDAO;
 import org.lenskit.inject.Shareable;
 import org.lenskit.inject.Transient;
+import org.lenskit.transform.normalize.UserVectorNormalizer;
 import org.lenskit.util.IdBox;
 import org.lenskit.util.collections.LongUtils;
 import org.lenskit.util.io.ObjectStream;
