+++ lenskit-knn/src/test/java/org/lenskit/knn/item/ItemItemRecommenderTest.java
@@ -23,26 +23,19 @@ package org.lenskit.knn.item;
 import it.unimi.dsi.fastutil.longs.LongArrayList;
 import it.unimi.dsi.fastutil.longs.LongOpenHashSet;
 import org.junit.After;
-import org.lenskit.api.RecommenderBuildException;
-import org.lenskit.LenskitConfiguration;
-import org.lenskit.data.dao.DataAccessException;
-import org.lenskit.data.dao.DataAccessObject;
-import org.lenskit.data.dao.EventCollectionDAO;
-import org.lenskit.data.dao.EventDAO;
-import org.lenskit.data.dao.file.StaticDataSource;
-import org.lenskit.data.ratings.Rating;
-import org.grouplens.lenskit.transform.normalize.DefaultUserVectorNormalizer;
-import org.grouplens.lenskit.transform.normalize.IdentityVectorNormalizer;
-import org.grouplens.lenskit.transform.normalize.UserVectorNormalizer;
-import org.grouplens.lenskit.transform.normalize.VectorNormalizer;
 import org.junit.Before;
 import org.junit.Test;
+import org.lenskit.LenskitConfiguration;
 import org.lenskit.LenskitRecommender;
 import org.lenskit.LenskitRecommenderEngine;
-import org.lenskit.api.ItemRecommender;
-import org.lenskit.api.ItemScorer;
-import org.lenskit.api.Result;
-import org.lenskit.api.ResultMap;
+import org.lenskit.api.*;
+import org.lenskit.data.dao.DataAccessObject;
+import org.lenskit.data.dao.file.StaticDataSource;
+import org.lenskit.data.ratings.Rating;
+import org.lenskit.transform.normalize.DefaultUserVectorNormalizer;
+import org.lenskit.transform.normalize.IdentityVectorNormalizer;
+import org.lenskit.transform.normalize.UserVectorNormalizer;
+import org.lenskit.transform.normalize.VectorNormalizer;
 
 import java.util.ArrayList;
 import java.util.List;
